+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.67.7
 
+- Swift: Allow UnidentifiedSenderMessageContent to be constructed from a message type and opaque bytes. (Contributed by @saman3d!)