+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
 v0.67.7
 
 - Swift: Allow UnidentifiedSenderMessageContent to be constructed from a message type and opaque bytes. (Contributed by @saman3d!)
+
+- net: Add a client for the registration verification service. This is currently
+  only available via the Node bindings.